requirements.txt
attrs~=23.1.0
certifi~=2023.7.22
charset-normalizer==2.0.8
greenlet~=2.0.2
idna~=3.4
iniconfig~=2.0.0
packaging~=23.1
playwright~=1.38.0
pluggy==1.0.0
py==1.11.0
pyee~=9.0.4
pyparsing~=3.1.1
pytest~=7.3.2
pytest-base-url==1.4.2
pytest-playwright==0.2.3
python-slugify==5.0.2
requests==2.26.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
urllib3~=2.0.4
websockets==10.1
colorama~=0.4.6
pip~=23.2.1
configparser~=5.3.0
wheel~=0.38.4
Pillow~=10.0.0
Jinja2~=3.1.2
trio~=0.22.2
pytz~=2023.3.post1
PyYAML~=6.0.1
MarkupSafe~=2.1.3
numpy~=1.25.2
setuptools~=68.0.0
cffi~=1.15.1
pyinstaller~=5.13.2
six~=1.16.0
pandas~=2.0.3
lxml~=4.9.3
python-dateutil~=2.8.2
PyMySQL~=1.1.0
python-dotenv~=1.0.0